feat: CE-1455 (#1055)
