+++ backend/src/types/models/case-files/assessment/assessment-action.ts
@@ -3,7 +3,7 @@ export class AssessmentActionDto {
   actor: UUID;
   date: Date;
   actionCode: string;
-  shortDescription: string;
-  longDescription: string;
+  shortDescription?: string;
+  longDescription?: string;
   activeIndicator: boolean;
 }
