+++ backend/src/types/models/case-files/assessment/create-assessment-input.ts
@@ -1,0 +1,6 @@
+import { BaseCaseFileInput } from "../base-case-file-input";
+import { AssessmentDto } from "./assessment";
+
+export interface CreateAssessmentInput extends BaseCaseFileInput {
+  assessment: AssessmentDto;
+}
