+++ frontend/cypress/e2e/allegation-details-create.cy.ts
@@ -19,7 +19,7 @@ describe("Complaint Create Page spec - Create View", () => {
     methodComplaintReceived: "RAPP",
     status: "Closed",
     statusIndex: 1,
-    assigned: "Kot, Steve",
+    assigned: "TestAcct, ENV",
     assignedIndex: 1,
     violationInProgressIndex: 1,
     violationInProgressString: "No",
