+++ frontend/cypress/e2e/change-complaint-status-details-view.cy.ts
@@ -21,6 +21,12 @@ describe("Complaint Change Status spec - Details View", () => {
       actionRequired: "Yes",
       toastText: "Assessment has been saved",
     };
+    // If assessment button exists, click it
+    cy.get("#outcome-assessments").then(($assessmentSection) => {
+      if ($assessmentSection.find("#outcome-report-add-assessment").length) {
+        cy.get("#outcome-report-add-assessment").click({ force: true });
+      }
+    });
     cy.get(".comp-outcome-report-complaint-assessment").then(function ($assessment) {
       if ($assessment.find("#outcome-save-button").length) {
         cy.fillInHWCSection(sectionParams).then(() => {
@@ -43,6 +49,7 @@ describe("Complaint Change Status spec - Details View", () => {
         cy.navigateToDetailsScreen(COMPLAINT_TYPES.HWCR, "23-000076", true);
         cy.assignSelfToComplaint();
         fillInAssessmentSection();
+        cy.waitForSpinner();
       } else {
         cy.navigateToDetailsScreen(COMPLAINT_TYPES.ERS, "23-006888", true);
         cy.assignSelfToComplaint();
