+++ frontend/cypress/e2e/hwcr-details-create-enter-coordinates.cy.ts
@@ -24,7 +24,7 @@ describe("Complaint Create Page spec - Enter Coordinates - Create View", () => {
     speciesIndex: 3,
     status: "Closed",
     statusIndex: 1,
-    assigned: "Kot, Steve",
+    assigned: "TestAcct, ENV",
     assignedIndex: 1,
   };
