+++ frontend/cypress/e2e/hwcr-details-edit.cy.ts
@@ -32,7 +32,7 @@ describe("Complaint Edit Page spec - Edit View", () => {
     speciesIndex: 1,
     status: "Open",
     statusIndex: 0,
-    assigned: "Kot, Steve",
+    assigned: "TestAcct, ENV",
     assignedIndex: 0,
   };
 
@@ -78,7 +78,7 @@ describe("Complaint Edit Page spec - Edit View", () => {
     speciesIndex: 3,
     status: "Closed",
     statusIndex: 1,
-    assigned: "Kot, Steve",
+    assigned: "TestAcct, ENV",
     assignedIndex: 1,
   };
