+++ frontend/cypress/e2e/hwcr-outcome-assessment.cy.ts
@@ -12,6 +12,11 @@ describe("HWCR Outcome Assessments", () => {
     //This is required to make the tests re-runnable.  It's not great because it means it will only run the first time.
     //If we ever get the ability to remove an assessment this test suite should be rewritten to remove this conditional
     //and to add a test at the end to delete the assessment.
+    cy.get("#outcome-assessments").then(($assessmentSection) => {
+      if ($assessmentSection.find("#outcome-report-add-assessment").length) {
+        cy.get("#outcome-report-add-assessment").click({ force: true });
+      }
+    });
     cy.get(".comp-outcome-report-complaint-assessment").then(function ($assessment) {
       if ($assessment.find("#outcome-save-button").length) {
         cy.validateComplaint("23-033066", "Coyote");
@@ -20,12 +25,11 @@ describe("HWCR Outcome Assessments", () => {
         cy.get("#outcome-save-button").click();
 
         //validate error message
-        cy.get(".error-message").should(($error) => {
-          expect($error).to.contain.text("Required");
-        });
-
-        let inputs = ["#action-required-div", "#outcome-officer-div", "#assessment-checkbox-div"];
-        cy.hasErrorMessage(inputs, "Errors in form");
+        cy.get(".error-message")
+          .should("contain.text", "Required")
+          .should(($error) => {
+            expect($error).to.contain.text("Required");
+          });
       } else {
         cy.log("Test was previously run. Skip the Test");
         this.skip();
@@ -39,6 +43,11 @@ describe("HWCR Outcome Assessments", () => {
     //This is required to make the tests re-runnable.  It's not great because it means it will only run the first time.
     //If we ever get the ability to remove an assessment this test suite should be rewritten to remove this conditional
     //and to add a test at the end to delete the assessment.
+    cy.get("#outcome-assessments").then(($assessmentSection) => {
+      if ($assessmentSection.find("#outcome-report-add-assessment").length) {
+        cy.get("#outcome-report-add-assessment").click({ force: true });
+      }
+    });
     cy.get(".comp-outcome-report-complaint-assessment").then(function ($assessment) {
       if ($assessment.find("#outcome-save-button").length) {
         cy.validateComplaint("23-033066", "Coyote");
@@ -69,6 +78,11 @@ describe("HWCR Outcome Assessments", () => {
 
     cy.validateComplaint("23-033066", "Coyote");
 
+    cy.get("#outcome-assessments").then(($assessmentSection) => {
+      if ($assessmentSection.find("#outcome-report-add-assessment").length) {
+        cy.get("#outcome-report-add-assessment").click({ force: true });
+      }
+    });
     cy.get(".comp-outcome-report-complaint-assessment").then(function ($assessment) {
       if ($assessment.find("#assessment-edit-button").length) {
         cy.get("#assessment-edit-button").click();
@@ -106,7 +120,7 @@ describe("HWCR Outcome Assessments", () => {
 
     cy.validateComplaint("23-033066", "Coyote");
 
-    cy.get(".comp-outcome-report-complaint-assessment").then(function ($assessment) {
+    cy.get("#outcome-assessments").then(function ($assessment) {
       if ($assessment.find("#assessment-edit-button").length) {
         cy.get("#assessment-edit-button").click();
