+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "nr-sample-natcomplaints",
-  "version": "1.9.5",
+  "version": "1.9.6",
   "private": true,
   "dependencies": {
     "@craco/craco": "^7.1.0",
