+++ frontend/src/app/components/containers/complaints/outcomes/hwcr-assessment/hwcr-assessment-link-complaint-search.tsx
@@ -62,7 +62,7 @@ const CustomHint = ({ children, className, hintText }: HintProps) => {
   );
 };
 
-export const HWCRComplaintAssessmentLinkComplaintSearch: FC<Props> = ({
+export const HWCRAssessmentLinkComplaintSearch: FC<Props> = ({
   id = "linkedComplaint",
   onChange = () => {},
   errorMessage = "",
