+++ frontend/src/app/components/containers/complaints/outcomes/hwcr-outcome-report.tsx
@@ -1,5 +1,5 @@
 import { FC, useEffect } from "react";
-import { HWCRComplaintAssessment } from "./hwcr-complaint-assessment";
+import { HWCRAssessments } from "./hwcr-assessment/hwcr-assessments";
 import { HWCREquipment } from "./hwcr-equipment";
 import { HWCRFileReview } from "./hwcr-file-review";
 import { Notes } from "./notes";
@@ -28,7 +28,7 @@ export const HWCROutcomeReport: FC = () => {
       <div className="comp-details-section-header">
         <h2>Outcome report</h2>
       </div>
-      <HWCRComplaintAssessment id={id} />
+      <HWCRAssessments />
       <HWCRComplaintPrevention />
       <HWCREquipment />
       <HWCROutcomeByAnimalv2 />
