+++ frontend/src/app/types/app/case-files/assessment/assessment-action.ts
@@ -3,8 +3,8 @@ export interface AssessmentActionDto {
   actor: UUID;
   date: Date;
   actionCode: string;
-  shortDescription: string;
-  longDescription: string;
+  shortDescription?: string;
+  longDescription?: string;
   activeIndicator: boolean;
   isLegacy?: boolean;
 }
