+++ frontend/src/app/types/app/case-files/assessment/assessment.ts
@@ -1,13 +1,15 @@
 import { AssessmentActionDto } from "./assessment-action";
-import KeyValuePair from "../key-value-pair";
-export interface AssessmentDetailsDto {
+import KeyValuePair from "../../key-value-pair";
+export interface AssessmentDto {
+  id?: string;
+  agencyCode?: string;
   actionNotRequired: boolean;
   actionCloseComplaint: boolean;
   actionLinkedComplaintIdentifier: string;
   actionJustificationCode: string;
-  actionJustificationShortDescription: string;
-  actionJustificationLongDescription: string;
-  actionJustificationActiveIndicator: boolean;
+  actionJustificationShortDescription?: string;
+  actionJustificationLongDescription?: string;
+  actionJustificationActiveIndicator?: boolean;
   actions: Array<AssessmentActionDto>;
   contactedComplainant: boolean;
   attended: boolean;
