+++ frontend/src/app/types/app/case-files/assessment/create-assessment-input.ts
@@ -1,0 +1,12 @@
+import { UUID } from "crypto";
+import { AssessmentDto } from "./assessment";
+export interface CreateAssessmentInput {
+  leadIdentifier: string;
+  caseIdentifier?: string;
+  agencyCode: string;
+  caseCode: string;
+  actor: UUID;
+  createUserId: string;
+  updateUserId: string;
+  assessment: AssessmentDto;
+}
