+++ frontend/src/app/types/app/case-files/assessment/update-assessment-input.ts
@@ -1,0 +1,9 @@
+import { AssessmentDto } from "./assessment";
+export interface UpdateAssessmentInput {
+  leadIdentifier: string;
+  caseIdentifier: string;
+  agencyCode: string;
+  caseCode: string;
+  updateUserId: string;
+  assessment: AssessmentDto;
+}
