--- frontend/src/app/types/app/case-files/create-assessment-input.ts
@@ -1,4 +1,0 @@
-import { CaseFileDto } from "./case-file";
-export interface CreateAssessmentInput {
-  createAssessmentInput: CaseFileDto;
-}
