--- frontend/src/app/types/app/case-files/update-assessment-input.ts
@@ -1,4 +1,0 @@
-import { CaseFileDto } from "./case-file";
-export interface UpdateAssessmentInput {
-  updateAssessmentInput: CaseFileDto;
-}
