+++ webeoc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "webeoc",
-  "version": "1.9.5",
+  "version": "1.9.6",
   "description": "",
   "author": "",
   "private": true,
